refactor(domain): move AdvisoryDatabaseInterface to Domain/Port
The interface is documented as a public extension point (host applications
implement it and alias their service to plug in custom CVE feeds), so its
canonical home is the Domain port namespace alongside LLMClientInterface,
AttackerCacheInterface, etc. Relocating it before 1.0 avoids a future BC
break — the previous Infrastructure FQCN was load-bearing.

Concrete adapters (ComposerAuditAdvisoryDatabase, InMemoryAdvisoryDatabase)
stay under Audit\Infrastructure\Advisory\. Host applications aliasing the
interface in config/services.yaml must update the FQCN; the simplified
listing in docs/versioning.md records this.
